separted fill functions
